style: Move grid-auto-flow outside of mako
